pyproject.toml
[tool.poetry]
name = "orders-service"
version = "0.1.0"
description = "Simple orders service"
authors = ["Mohammed Mwijaa <mm.mwijaa@gmail.com>"]
license = "GNU General Public Licence v3"

[tool.poetry.dependencies]
python = "^3.8"
drf-oidc-auth = "^1.0.0"
psycopg2-binary = "^2.8.6"
celery = "^5.0.5"
redis = "^3.5.3"
python-decouple = "^3.4"
dj-database-url = "^0.5.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
pytest-cov = "^2.11.1"
requests = "^2.25.1"
flake8 = "^3.8.4"
black = "^20.8b1"
pre-commit = "^2.10.1"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-raq --cov=. --cov-report xml"
testpaths = [
	"tests"
]

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
